requirements.txt
praw
buttplug-py
